A simple re-factor to de-couple Calendar Parsing from MMM
* My calendars were not behaving as 'expected' so to more easily test them
  I re-factored the code lightly so I could create a simple test harness and
  debug the problem more effectively.
